Alpha contributor design
