+++ MovementCohort1/MovementCohort1/App/Assets.xcassets/khadia-header.imageset/KHADIA.png
